Moved logging setup into app.py
Modified Ticker to accept one end of a Pipe so
that it can communicate with it's parent process.

Created TheLoop class which will control child processes and
handling of interprocess communication.

Removed recursive calls to ticker methods when errors occur. Instead
we send a message to TheLoop to decide if it wants to fork a new
child process.

Some changes to log messages.

Added TheLoop unit test. Currently only tests loop method to ensure
appropriate handlers get called.

Modified Ticker tests to jive with passed in Pipe and the fact that
it no longer recursively calls 'start'.